+++ include/petrush/petrush.h
@@ -6,7 +6,7 @@
 #ifndef PETRUSH_PETRUSH_H
 #define PETRUSH_PETRUSH_H
 
-#define PETRUSH_VERSION "0.1.0"  // S0 - Onda 1 Gate achieved
+#define PETRUSH_VERSION "0.2.0"  // Onda 3: pipes + redirecionamento
 #define PETRUSH_NAME "petrush"
 
 #endif /* PETRUSH_PETRUSH_H */